Fix dtype mismatch in evolutionary step tensor creation
Specify dtype=torch.float32 when creating tensors from evolutionary
offspring parameters to match the model's expected dtype and prevent
RuntimeError.

Fixes crash: "mat1 and mat2 must have the same dtype, but got Double and Float"

+++ models/trainer.py
@@ -352,8 +352,8 @@ class HybridTrainer:
                         mesh = self.renderer.render_to_mesh(scad_code)
                         metrics = self.evaluator.evaluate(mesh, params)
                         self.buffer.add(params, metrics)
-                        
-                        params_tensor = torch.tensor(params.to_vector()).to(self.device)
+
+                        params_tensor = torch.tensor(params.to_vector(), dtype=torch.float32).to(self.device)
                         evo_results.append((params, metrics, params_tensor))
                         
                     except Exception as e:
